pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guide3d"
version = "0.1.dev1"
description = ""
license = { file = "LICENSE" }
authors = [
    {name = "Tudor Jianu", email = "tudorjnu@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Free for non-commercial use",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python",
    "numpy<2",
    "matplotlib",
    "pytorch-lightning",
    "torchvision",
    "wandb",
    "scipy",
    "trimesh",
    "spectra",
    "seaborn",
    "lxml",
    "shapely",
    "plotly",
    "transformers",
    "colour",
    "gdown"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
